Fix manual refactoring and attempt threads
Fixed the manual refactoring so that the drives would actually go away and the radio buttons would be reduced when a drive is removed. Also, threading was attempted but not completed due to Qt not allowing GUI functions outside the main thread. This threading will allow the drives to continuously update and *hopefully* wipe.